requirements.txt
git+https://github.com/facebookresearch/segment-anything.git
regex==2023.12.25
tqdm==4.66.4
diffusers==0.27.2
transformers==4.40.1
accelerate==0.30.0
scipy==1.11.4
xformers==0.0.25
opencv-python==4.8.0.76
pycocotools==2.0.7
matplotlib==3.7.1
onnxruntime==1.17.3
onnx==1.16.0
gradio==4.29.0
ipython==7.34.0
numpy==1.25.2
pillow==9.4.0
segment_anything